+++ src/index.ts
@@ -161,6 +161,7 @@ export * from "./components/CashflowChart";
 export * from "./components/DnDHierarchySortBar";
 export * from "./components/ScrubChart";
 export * from "./components/CashflowScrubChart";
+export * from "./components/WorkProgressCard";
 
 // Styles - import this in your app: import "solid-ui-components/styles.css"